requirements.txt
################################################################################
# This requirements file has been automatically generated from `Pipfile` with
# `pipenv-to-requirements`
#
#
# This has been done to maintain backward compatibility with tools and services
# that do not support `Pipfile` yet.
#
# Do NOT edit it directly, use `pipenv install [-d]` to modify `Pipfile` and
# `Pipfile.lock` and then regenerate `requirements*.txt`.
################################################################################

aiosmtplib==1.1.3
azure-functions-worker==1.1.2
azure-functions==1.2.1
grpcio-tools==1.26.0
grpcio==1.26.0
marshmallow==3.6.1
protobuf==3.12.2
six==1.15.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sremail==1.0.0
